Force number to string conversion
- The problem
It seems that there is a problem with type mismatch on Cosmic.
The issue is that if you put a the number '10' in a textfield, then the response is converted to a `number`, but if you put the number '07', then it gets converted to a `string`.

- Workaround
convert every number to a string, which can remove this type mismatch issue